Backend/requirements.txt
firebase-admin
flask
flask-cors
openai
numpy
pyyaml
requests
websockets
# chromadb
# torch
# gpt4all
# transformers